requirements.txt
Django==1.5.1
PyYAML==3.10
argparse==1.2.1
distribute==0.6.34
dj-database-url==0.2.1
django-appconf==0.6
django-compressor==1.3
django-configurations==0.3
django-cors-headers==0.11
django-tastypie==0.9.15
gunicorn==0.17.4
mimeparse==0.1.3
psycopg2==2.5.1
python-dateutil==2.1
six==1.3.0
wsgiref==0.1.2
